+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.5.0 - UNRELEASED
+
+### Added
+
+- Event calendar button to display current or upcoming theme
+
+### Removed
+
+- Option to disable particles. The performance issues have been resolved. (simplifies settings screen)
+
 ## Version 1.4.0 - 2024-10-06
 
 ### Added